# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEET Automotive Community API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete user onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingRequest"
      responses:
        "200":
          description: User profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get list of events
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete/cancel event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [viewer, participant, organizer]
        organizationName:
          type: string
          nullable: true
        contactLink:
          type: string
          nullable: true
        viewerSilhouette:
          type: string
          nullable: true
          enum: [bicycle, scooter, skateboard, cart]
        interestCategories:
          type: array
          items:
            type: string
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - telegramId
        - username
        - displayName
        - role
        - onboardingComplete
        - interestCategories
        - createdAt
    OnboardingRequest:
      type: object
      properties:
        telegramId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [viewer, participant, organizer]
        viewerSilhouette:
          type: string
          nullable: true
          enum: [bicycle, scooter, skateboard, cart]
        organizationName:
          type: string
          nullable: true
        contactLink:
          type: string
          nullable: true
        interestCategories:
          type: array
          items:
            type: string
      required:
        - telegramId
        - username
        - displayName
        - role
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [motorsport, exhibition, cruise, club]
        subcategories:
          type: array
          items:
            type: string
        organizerId:
          type: integer
        organizerName:
          type: string
        organizerContact:
          type: string
          nullable: true
        date:
          type: string
        endDate:
          type: string
          nullable: true
        location:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        maxParticipants:
          type: integer
          nullable: true
        isPrivate:
          type: boolean
        coverImageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [upcoming, ongoing, finished, cancelled]
        applicationsCount:
          type: integer
        viewersCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - subcategories
        - organizerId
        - organizerName
        - date
        - location
        - isPrivate
        - status
        - applicationsCount
        - viewersCount
        - createdAt
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [motorsport, exhibition, cruise, club]
        subcategories:
          type: array
          items:
            type: string
        date:
          type: string
        endDate:
          type: string
          nullable: true
        location:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        maxParticipants:
          type: integer
          nullable: true
        isPrivate:
          type: boolean
        coverImageUrl:
          type: string
          nullable: true
      required:
        - title
        - category
        - subcategories
        - date
        - location
        - isPrivate
...