# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MEET Automotive Community API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserViewerSilhouette } from "./userViewerSilhouette";

export interface User {
  id: number;
  telegramId: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  role: UserRole;
  organizationName?: string | null;
  contactLink?: string | null;
  viewerSilhouette?: UserViewerSilhouette;
  interestCategories: string[];
  onboardingComplete: boolean;
  createdAt: string;
}